+++ src/ff_util.h
@@ -47,7 +47,7 @@ void file_menu_close_dir(void);
 uint16_t file_menu_get_size(void);
 void file_menu_full_sort(void);
 void file_menu_sort_entry(uint16_t scope_start, uint16_t scope_end_1);
-FRESULT file_menu_get_fname(uint16_t order, char *str, uint16_t size);
+FRESULT file_menu_get_fname(uint16_t order, char *str, size_t size);
 FRESULT file_menu_get_obj(uint16_t order, FsBaseFile *file);
 int file_menu_is_dir(uint16_t order);
 void file_menu_idle(void);
